// src/models/CarInterfaces.ts
export interface Car {
    "id": number
    "name": string
    "model": string
    "year": number
    "color": string
    "price": number
    "latitude": number
    "longitude": number
}

export interface UpdatedCar {
    "name": string
    "model": string
    "year": number
    "color": string
    "price": number
}